feat : construct classes

+++ app/src/main/java/com/example/cee_project1/data/InvestOption.kt
@@ -3,7 +3,7 @@ package com.example.cee_project1.data
 data class InvestOption(
     var name : String = "",
     var amount : Int = 0,
-    var value : Int = 0,
+    var value : Double = 0.0,
     var price : Int = 0,
     var maxPoint : Int = 0,
     var minPoint : Int = 0
